+++ src/db/mysql/MySqlAPI.h
@@ -288,7 +288,7 @@ public:
 
     virtual double getSuccessRate(std::string source, std::string destination);
 
-    virtual int getAvgThroughput(std::string source, std::string destination, int activeTransfers);
+    virtual double getAvgThroughput(std::string source, std::string destination, int activeTransfers);
 
     virtual void updateProtocol(const std::string& jobId, int fileId, int nostreams, int timeout, int buffersize, double filesize);
